test: move cmd/main_test.go to pkg/github/handlers_test.go
This commit moves the actionconfig and its unit-tests into the package
`pkg`. This fixes the unit-test execution that are currently failing due
to issues with flags in `cmd`.
